+++ ProWin/Sources/Engine/GraphicsBridge.h
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+
+@interface GraphicsBridge : NSObject
+
++ (instancetype)sharedInstance;
+- (void)presentFrame;
+- (void*)getMetalDevice;
+
+@end